+++ tools/builtins/grep.rkt
@@ -5,7 +5,7 @@
          racket/file
          racket/list
          racket/path
-         (only-in "../tool.rkt" make-success-result make-error-result)
+         (only-in "../tool.rkt" make-success-result make-error-result tool-result?)
          (only-in "builtin-helpers.rkt" require-safe-path!)
          (only-in "../../util/glob.rkt" glob->regexp)
          (only-in "../../util/path-helpers.rkt"
@@ -15,7 +15,7 @@
          (only-in "../../util/path-filters.rkt" hidden-name? should-skip-entry? skip-dirs))
 
 (require racket/contract)
-(provide (contract-out [tool-grep (-> hash? any/c)]))
+(provide (contract-out [tool-grep (-> hash? tool-result?)]))
 
 ;; ============================================================
 ;; Defaults